+++ frontend/src/containers/App.tsx
@@ -1,7 +1,9 @@
+import HomePage from "../screens/HomePage";
+
 export default function App() {
   return (
     <div className="App">
-      TakeSurvey
+      <HomePage />
     </div>
   );
 }
