+++ frontend/src/utils/react-app-env.d.ts
